wildfire-backend/requirements.txt
fastapi
uvicorn
httpx
python-multipart
numpy
shapely
pydantic
python-dotenv
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
bcrypt
email-validator
passlib
cloudinary
